Split up utility functions into separate files.
